7. Modularize the code 7. Modularize the code final
