+++ staff/gustavo-rojas/course-3/task-2/scheduler/src/components/CourseList.js
@@ -1,5 +1,6 @@
 import React, { useState } from 'react';
-import { hasConflict,getCourseTerm,getCourseNumber,terms } from '/mnt/c/ubiqum/ubiqum-bootcamp/staff/gustavo-rojas/course-3/task-2/scheduler/src/utilities/times.js';
+import { getCourseTerm,terms } from '/mnt/c/ubiqum/ubiqum-bootcamp/staff/gustavo-rojas/course-3/task-2/scheduler/src/utilities/times.js';
+import Course from '/mnt/c/ubiqum/ubiqum-bootcamp/staff/gustavo-rojas/course-3/task-2/scheduler/src/components/Course.js';
 const CourseList = ({ courses }) => {
     const [term, setTerm] = useState('Fall');
     const [selected, setSelected] = useState([]);
@@ -20,9 +21,6 @@ const CourseList = ({ courses }) => {
     );
   };
 
-  const toggle = (x, lst) => (
-    lst.includes(x) ? lst.filter(y => y !== x) : [x, ...lst]
-  );
 
   const TermButton = ({term, setTerm, checked}) => (// We neeed change className for class  in this code class="btn btn-success m-1 p-2" htmlFor={term}>
   <>
@@ -34,23 +32,7 @@ const CourseList = ({ courses }) => {
   </>
 );
 
-const Course = ({ course, selected, setSelected }) => {
-  const isSelected = selected.includes(course);
-  const isDisabled = !isSelected && hasConflict(course, selected);
-  const style = {
-    backgroundColor: isDisabled? 'lightgrey' : isSelected ? 'lightgreen' : 'white'
-  };
-  return (
-    <div className="card m-1 p-2" 
-      style={style}
-      onClick={isDisabled ? null : () =>  setSelected(toggle(course, selected))}>
-      <div className="card-body">
-        <div className="card-title">{ getCourseTerm(course) } CS { getCourseNumber(course) }</div>
-        <div className="card-text">{ course.title }</div>
-      </div>
-    </div>
-  );
-};
+
 const TermSelector = ({term, setTerm}) => (
     <div className="btn-group">
     {
